refactoring runmat-wasm for maintainability (+ wasm clippy now green)
