+++ crates/runmat-wasm/src/api/mod.rs
@@ -1,0 +1,5 @@
+// The stable wasm-bindgen boundary exposed to JavaScript.
+pub(crate) mod init;
+pub(crate) mod plot;
+pub(crate) mod session;
+pub(crate) mod streams;
