+++ crates/runmat-wasm/src/runtime/filesystem/mod.rs
@@ -1,0 +1,6 @@
+mod bindings;
+mod convert;
+mod handle;
+mod provider;
+
+pub(crate) use provider::install_js_fs_provider;
