+++ crates/runmat-wasm/src/runtime/mod.rs
@@ -1,0 +1,7 @@
+// Shared implementation support consumed by the wasm API layer.
+pub(crate) mod config;
+pub(crate) mod filesystem;
+pub(crate) mod gpu;
+pub(crate) mod logging;
+pub(crate) mod snapshot;
+pub(crate) mod state;
