+++ crates/runmat-wasm/src/wire/mod.rs
@@ -1,0 +1,4 @@
+// Rust-to-JS wire contracts and serialization helpers.
+pub(crate) mod errors;
+pub(crate) mod payloads;
+pub(crate) mod value;
