+++ crates/runmat-wasm/src/wire/value.rs
@@ -1,0 +1,245 @@
+use runmat_builtins::{NumericDType, ObjectInstance, StructValue, Value};
+use runmat_core::{matlab_class_name, value_shape};
+use serde_json::{json, Map as JsonMap, Value as JsonValue};
+
+pub(crate) const MAX_DATA_PREVIEW: usize = 4096;
+const MAX_STRUCT_FIELDS: usize = 64;
+const MAX_OBJECT_FIELDS: usize = 64;
+const MAX_OUTPUT_LIST_ITEMS: usize = 64;
+const MAX_RECURSION_DEPTH: usize = 2;
+
+pub(crate) fn value_to_json(value: &Value, depth: usize) -> JsonValue {
+    if depth >= MAX_RECURSION_DEPTH {
+        return json!({
+            "kind": "display",
+            "className": matlab_class_name(value),
+            "shape": value_shape(value),
+            "value": value.to_string(),
+        });
+    }
+
+    match value {
+        Value::Int(iv) => json!({
+            "kind": "int",
+            "className": iv.class_name(),
+            "value": iv.to_i64(),
+            "shape": scalar_shape(),
+        }),
+        Value::Num(n) => json!({
+            "kind": "double",
+            "value": n,
+            "shape": scalar_shape(),
+        }),
+        Value::Complex(re, im) => json!({
+            "kind": "complex",
+            "real": re,
+            "imag": im,
+            "shape": scalar_shape(),
+        }),
+        Value::Bool(b) => json!({
+            "kind": "logical",
+            "value": b,
+            "shape": scalar_shape(),
+        }),
+        Value::LogicalArray(arr) => {
+            let (preview, truncated) = preview_slice(&arr.data, MAX_DATA_PREVIEW);
+            let rows = arr.shape.first().copied().unwrap_or(0);
+            let cols = arr.shape.get(1).copied().unwrap_or(0);
+            json!({
+                "kind": "logical-array",
+                "shape": arr.shape,
+                "rows": rows,
+                "cols": cols,
+                "preview": preview.iter().map(|v| *v != 0).collect::<Vec<bool>>(),
+                "length": arr.data.len(),
+                "truncated": truncated,
+            })
+        }
+        Value::String(s) => json!({
+            "kind": "string",
+            "value": s,
+            "shape": vec![1, s.chars().count()],
+        }),
+        Value::StringArray(sa) => {
+            let (preview, truncated) = preview_slice(&sa.data, MAX_DATA_PREVIEW);
+            json!({
+                "kind": "string-array",
+                "shape": sa.shape,
+                "rows": sa.rows,
+                "cols": sa.cols,
+                "preview": preview,
+                "length": sa.data.len(),
+                "truncated": truncated,
+            })
+        }
+        Value::CharArray(ca) => {
+            let s: String = ca.data.iter().copied().collect();
+            json!({
+                "kind": "char-array",
+                "rows": ca.rows,
+                "cols": ca.cols,
+                "shape": vec![ca.rows, ca.cols],
+                "text": s,
+            })
+        }
+        Value::Tensor(t) => {
+            let (preview, truncated) = preview_slice(&t.data, MAX_DATA_PREVIEW);
+            json!({
+                "kind": "tensor",
+                "shape": t.shape,
+                "rows": t.rows,
+                "cols": t.cols,
+                "dtype": match t.dtype {
+                    NumericDType::F32 => "single",
+                    NumericDType::F64 => "double",
+                },
+                "preview": preview,
+                "length": t.data.len(),
+                "truncated": truncated,
+            })
+        }
+        Value::ComplexTensor(t) => {
+            let (preview, truncated) = preview_slice(&t.data, MAX_DATA_PREVIEW);
+            let preview: Vec<JsonValue> = preview
+                .into_iter()
+                .map(|(re, im)| json!({ "real": re, "imag": im }))
+                .collect();
+            json!({
+                "kind": "complex-tensor",
+                "shape": t.shape,
+                "rows": t.rows,
+                "cols": t.cols,
+                "preview": preview,
+                "length": t.data.len(),
+                "truncated": truncated,
+            })
+        }
+        Value::Cell(ca) => json!({
+            "kind": "cell",
+            "shape": ca.shape,
+            "rows": ca.rows,
+            "cols": ca.cols,
+            "length": ca.data.len(),
+        }),
+        Value::OutputList(values) => {
+            let truncated = values.len() > MAX_OUTPUT_LIST_ITEMS;
+            let items: Vec<JsonValue> = values
+                .iter()
+                .take(MAX_OUTPUT_LIST_ITEMS)
+                .map(|v| value_to_json(v, depth + 1))
+                .collect();
+            json!({
+                "kind": "output-list",
+                "length": values.len(),
+                "items": items,
+                "truncated": truncated,
+            })
+        }
+        Value::Struct(st) => struct_to_json(st, depth + 1),
+        Value::GpuTensor(handle) => {
+            let (rows, cols) = rows_cols_from_shape(&handle.shape);
+            json!({
+                "kind": "gpu-tensor",
+                "shape": handle.shape,
+                "rows": rows,
+                "cols": cols,
+                "deviceId": handle.device_id,
+                "bufferId": handle.buffer_id,
+            })
+        }
+        Value::Object(obj) => object_to_json(obj, depth + 1),
+        Value::HandleObject(handle) => json!({
+            "kind": "handle",
+            "className": handle.class_name,
+            "valid": handle.valid,
+        }),
+        Value::Listener(listener) => json!({
+            "kind": "listener",
+            "id": listener.id,
+            "event": listener.event_name,
+            "enabled": listener.enabled,
+            "valid": listener.valid,
+        }),
+        Value::FunctionHandle(name) => json!({
+            "kind": "function",
+            "name": name,
+        }),
+        Value::Closure(closure) => json!({
+            "kind": "closure",
+            "functionName": closure.function_name,
+            "captureCount": closure.captures.len(),
+        }),
+        Value::ClassRef(name) => json!({
+            "kind": "class-ref",
+            "name": name,
+        }),
+        Value::MException(ex) => json!({
+            "kind": "exception",
+            "identifier": ex.identifier,
+            "message": ex.message,
+            "stack": ex.stack,
+        }),
+    }
+}
+
+fn struct_to_json(st: &StructValue, depth: usize) -> JsonValue {
+    let mut fields = JsonMap::new();
+    let mut truncated = false;
+    for (idx, (name, field_value)) in st.fields.iter().enumerate() {
+        if idx >= MAX_STRUCT_FIELDS {
+            truncated = true;
+            break;
+        }
+        fields.insert(name.clone(), value_to_json(field_value, depth));
+    }
+    json!({
+        "kind": "struct",
+        "fieldOrder": st.field_names().take(MAX_STRUCT_FIELDS).cloned().collect::<Vec<_>>(),
+        "fields": fields,
+        "totalFields": st.fields.len(),
+        "truncated": truncated,
+    })
+}
+
+fn object_to_json(obj: &ObjectInstance, depth: usize) -> JsonValue {
+    let mut fields = JsonMap::new();
+    let mut truncated = false;
+    for (idx, (name, value)) in obj.properties.iter().enumerate() {
+        if idx >= MAX_OBJECT_FIELDS {
+            truncated = true;
+            break;
+        }
+        fields.insert(name.clone(), value_to_json(value, depth));
+    }
+    json!({
+        "kind": "object",
+        "className": obj.class_name,
+        "properties": fields,
+        "propertyCount": obj.properties.len(),
+        "truncated": truncated,
+    })
+}
+
+fn scalar_shape() -> Vec<usize> {
+    vec![1, 1]
+}
+
+fn rows_cols_from_shape(shape: &[usize]) -> (usize, usize) {
+    let rows = shape.first().copied().unwrap_or(0);
+    let cols = if shape.len() >= 2 {
+        shape[1]
+    } else if rows == 0 {
+        0
+    } else {
+        1
+    };
+    (rows, cols)
+}
+
+fn preview_slice<T: Clone>(data: &[T], limit: usize) -> (Vec<T>, bool) {
+    if data.len() > limit {
+        (data[..limit].to_vec(), true)
+    } else {
+        (data.to_vec(), false)
+    }
+}
